==> Pandas/Module 1/1. Data Structures- Series.ipynb <==
# Module 1: Data Structure - Series

from pandas import Series, DataFrame #Imported them in local name space(ofcourse will be used a lot)

import pandas as pd
import numpy as np

# Contents:
    1. Series
    2. Data_Frame

# 1. Series

One dimensional array

NumPy Data Type

Data labels called its index

obj = Series([1,2,4,5])

#INDEX- DATA format

obj

obj.values

obj.index #Default Index starts from zero to N-1

#Often its desirable to create any series with index:

obj2 = Series([4,9,8,7], index = ['a','b','c','d'])

obj2

Indexing Operations:

obj2['a']

obj2[obj2 > 4]

obj2*2 #Multiplication

np.exp(obj2)

Series can also be thought like a dictionary: For example it is mapping index value to data value.

'b' in obj2
'c' in obj2

'f' in obj2

#### Dictionary to Series Transformation: Here the key is recognised as index in series.

sdic = {'b':2000, 'c':3000, 'a':1000, 'd':4000}

ser = pd.Series(sdic) 

ser

#Proper arrangment:

alpha = ['a','b','c','d','e']

obj4 = Series(sdic, index = alpha) #assigning index with variable alpha created above

obj4

### isnull and #notnull will help us check if there is any missing data in our data file.

pd.isnull(obj4) 

pd.notnull(obj4)

#Some more operations:

obj3 = Series([10,20,30,40],index = ['a','b','c','d'])

obj3

obj4

obj3 + obj4

#### Both the Series object itself and its index have a name attribute

obj4.name = 'Money'

obj4.index.name = 'person'

obj4
