==> Pandas/Module 1/2. Data Frame- Basics- Creating New Data.ipynb <==
# DataFrame

#### A DataFrame represents a tabular, spreadsheet-like data structure containing an ordered collection of columns, each of which can be a different value type. It has both row and column index. It can be thought of as dictionary of series. Row oriented and column oriented operations are treated symmetrically. 

#### Data is stored as 1D or 2D blocks rather than list or dictionary. Higher dimensions are also possible using heirarchical structure. 

columns in data frame = series and colums + columns (i.e addition of series) = DF (Analogy)

#Constructing the DataFrame:

import pandas as  pd
from pandas import DataFrame

data = {'state' : ['Mah', 'UP', 'MP', 'Kar', 'TN', 'Naga'],
       'year': [1960, 1947, 1990, 1995, 2000, 1962],
        'pop': [1.1,1.2,1.6,1.1,1.0,2.0]}
frame = DataFrame(data)

We are using the pd.DataFrame() constructor to generate these DataFrame objects. The syntax for declaring a new one is a dictionary whose keys are the column names and whose values are a list of entries.

frame

The dictionary-list constructor assigns values to the column labels, but just uses an ascending count from 0 to n-1

We can also change the index labels:

pd.DataFrame({'state' : ['Mah', 'UP', 'MP', 'Kar', 'TN', 'Naga'],
       'year': [1960, 1947, 1990, 1995, 2000, 1962],
        'pop': [1.1,1.2,1.6,1.1,1.0,2.0]}, index = ['state1','state2','state3','state4','state5','state6'])

#Pass a column that is not appearing in data and it will return NaN

Recall that series does not have column name but have only overall name. While in Pandas we have index and column name

checking the dimensions:

frame.shape 

rows, columns = frame.shape

rows

columns

# Retrieving Columns using Indexing:

frame.columns

frame[['year','state']]

frame['year']

frame.year #use any one syntax

# Retriving rows:

frame.loc[frame['year']==1960]

frame.index = ['a','b','c','d','e','f']

frame.loc['a'] #Using indexing if we have specified any particular index to our frame
