==> Pandas/Module 1/3. Data Loading and Storage.ipynb <==
import pandas as pd
from pandas import DataFrame

csv means comma seperated values.

# Data Reading and Loading

data = pd.read_csv('/home/meta-sapien/Downloads/archive/football.csv')

data.head(5)

data.tail(5)

data.shape

shape attribute is used to check how big the data frame is 

To make pandas use column specified in csv file itself for the index (instead of creating a new one from scratch), we can specify an index_col

data = pd.read_csv('/home/meta-sapien/Downloads/archive/football.csv', index_col = 0)

data.head(5)

data = pd.read_csv('/home/meta-sapien/Downloads/archive/football.csv')

data.head(5)

data['Christiano Ronaldo'].max()

data['Neymar'].mean()

data['Neymar'].std() #standard deviation

data.describe()

# Data Storing

data.to_csv('foosball.csv')

ls #See that we are storing new file called foosball.csv in our default directory

# Data Loading

pd.read_csv('foosball.csv',header = None, names = ['a','b','c','d','e','f','g']).head(5) #However here new column is being added!!

pd.read_excel('/home/meta-sapien/Downloads/mlb2010.xls',sheet_name='pitch2010').head(5)

sheet_name = specifies the sheet name or also we can use sheet_name = [0,1,2] means 1st 3 sheets likewise

#by default it reads first sheet if not specified

data2 = pd.read_html('https://www.baseball-reference.com/players/s/semiema01.shtml#batting_standard')

type(data2)

Here the html gave us the list format so we can do some manipulation and change it into our required Dataframe format:

data2[0].to_csv('players.csv')

pd.read_csv('players.csv').head(5)
