==> Pandas/Module 1/4. Indexing in Pandas and Conditional Selection.ipynb <==
# INDEXING

Selecting specific values of a pandas DataFrame or Series to work on is an implicit step in almost any data operation

#Import the required Libraries
import pandas as pd
from pandas import DataFrame

#Create the DataFrame
data = pd.DataFrame({'Brand' : ['Maruti', 'Hyundai', 'Tata',
                                'Mahindra', 'Maruti', 'Hyundai',
                                'Renault', 'Tata', 'Maruti'],
                     'Year' : [2012, 2014, 2011, 2015, 2012, 
                               2016, 2014, 2018, 2019],
                     'Kms Driven' : [50000, 30000, 60000, 
                                     25000, 10000, 46000, 
                                     31000, 15000, 12000],
                     'City' : ['Gurgaon', 'Delhi', 'Mumbai', 
                               'Delhi', 'Mumbai', 'Delhi', 
                               'Mumbai','Chennai',  'Ghaziabad'],
                     'Mileage' :  [28, 27, 25, 26, 28, 
                                   29, 24, 21, 24]})

data

data['Brand'][0]

Notice that this is column first and row second type of indexing

# Index Based Selection : iloc

Selecting data based on its numerical position in the data. 

data.iloc[0]

#### Both loc and iloc are row-first, column-second.

data.iloc[0]['Brand']

This means that it's marginally easier to retrieve rows, and marginally harder to get retrieve columns

#Retriving the columns:

data.iloc[:,0:1]

data.iloc[[0,2,3],:]

negative numbers can be used in selection

This will start counting forwards from the end of the values

data.iloc[-5:]

# Label Based Selection- loc

it's the data index value, not its position, which matters

data.loc[0,'Brand']

iloc is conceptually simpler than loc because it ignores the dataset's indices. When we use iloc we treat the dataset like a big matrix (a list of lists), one that we have to index into by position. loc, by contrast, uses the information in the indices to do its work.

 When dataset usually has meaningful indices, it's usually easier to do things using loc instead

data.loc[:,['Brand','Year','City']]

#Choosing between loc and iloc - 

       |   loc                       | iloc
---------------------------------------------------------
Range  | indexing[0:10] >> 0,...,10  |indexing[0:10] >> 0,....,09

returns| 11 values                   | 10 values                  

# Manipulating the Index

data.set_index('Brand')

#Application: Now label based selection derives its power from the labels in the index. Critically the index we use in not immutable. We can manipulate the index in way we see fit

This is useful if you can come up with an index for the dataset which is better than the current one.

This usually returns new dataframe and does not modify the original so we need to reasign it:

Here the value inside set index acts like Hash Key

data_2 = data.set_index('Brand')

data_2

#to modify the original one:

data_2.set_index('City',inplace= True)

data_2

data_2.reset_index(inplace=True)
data_2

# Conditional Selection

data.Brand == 'Maruti'

data.loc[data.Brand == 'Maruti']

Now here we can easily see that this fuctionality can be used inside .loc 

#And Operation:

data.loc[(data.Brand == 'Maruti') & (data.Mileage == 28)]

#OR Operation

data.loc[(data.Brand == 'Maruti') | (data.Mileage == 28)]

#PANDAS Built in conditional Selectors: isin and isnull

The first is isin. isin is lets you select data whose value "is in" a list of values.

data.loc[data.Brand.isin(['Maruti','Tata'])]

The second is isnull (and its companion notnull). These methods let you highlight values which are (or are not) empty (NaN)

data.loc[data.Year.notnull()]

data.loc[data.Year.isnull()]

data.loc[9] = {'Brand':'Ford','Year': 2020,'kms Driven':200}

data

data.loc[data.City.isnull()]

#Now if we observe closely then our data has kms Driven written with space in between how to manipulate this then: Use the below syntax

data.loc[data['Kms Driven'].isnull()]

data['reverse'] = range(len(data),0,-1)

data

data.pop('reverse')

data

# Question: Give me a car with its Brand,Year and Mileage which has maximum kms driven

data[['Brand','Year','Mileage']][data['Kms Driven'] == data['Kms Driven'].max()]
