==> Pandas/Module 1/5. More on Reading and Writing.ipynb <==
# Reading

import pandas as pd
from pandas import DataFrame
data = pd.read_csv('cars.csv')

data

#Here we do not need the First row then-

data = pd.read_csv('cars.csv',skiprows=1)
data

data = pd.read_csv('cars.csv',header=1) #Does the same function as skip rows
data

index starts from zero that means the header = 1 means the first row is used as header

data1 = pd.read_csv('cars.csv', header = None, skiprows = 1, index_col = 0, names = ['Index','Brandd','Yearr','Kmss','Cityy','milaeagee'])
data1

data1 = pd.read_csv('cars.csv', header = None, skiprows = 2, index_col = 0, names = ['Index','Brandd','Yearr','Kmss','Cityy','milaeagee'])
data1

1. skiprows- skip the row
2. index_col = 0 
3. names=[] needs to be written if Header = 0

data = pd.read_csv('cars.csv',nrows = 3) #It will read only 3 rows
data

data = pd.read_csv('cars.csv',na_values=['Tata','Delhi'])
data

it is important to notice that na_value replaced the values given e.g. Delhi and Tata with NaN!!

Dictionary gives us more specific use case for converting some values to NaN

data = pd.read_csv('cars.csv',header=1,na_values={'Year':[2011,2012],
                                        'City':['Mumbai','Chennai']})
data

#As seen in the result we can convert specific values in particular columns into NaN values which is not possible in case of general list

# Writing

data.to_csv('new.csv', index = False, header = 1,columns =['Brand','Year','Kms Driven','City','Mileage'])

pd.read_csv('new.csv')

1. index = False will not write the index
2. header = 1 will select header as Brand and not the wrong row
3. colums = [  ] will write only those columns which are specified in the given list

data.to_excel('new.xlsx',sheet_name = 'carsdata',startrow = 1, startcol=1)


Two data frames and then how can we add them to single excel file?

# excel_writer:

with pd.ExcelWriter('new.xlsx') as writer:
    data.to_excel(writer,sheet_name = '001')
    data1.to_excel(writer,sheet_name = '002')

pd.read_excel('new.xlsx',sheet_name = '001').head(3)

pd.read_excel('new.xlsx',sheet_name = '002').head(3)

# Converter function:
####Helps in converting messy data into something useful

def convert_city_cell(cell):
    if cell == '':
        return 'Pune'
    return cell

data = pd.read_csv('new.csv',converters = {
    'City': convert_city_cell
})
data

#Look pandas read documentation for more functionalities like seperator and delimitor
