==> Pandas/Module 1/6. Summary Functions and Maps.ipynb <==
 The data does not always come out of memory in the format we want it. This notebook covers the operations that we need to doo just to make our data RIGHT!

import pandas as pd

data = pd.read_csv('new.csv')
data.head(3)

# Summary Functions:

data.Mileage.describe()

data.Brand.describe() #simple summary statistic

data.City.unique() #To see a list of unique values

data.City.value_counts() #how often they occur in the dataset

data.describe() #Summary Statistics

# MAPS

 In data science we often have a need for creating new representations from existing data, or for transforming data from the format it is in now to the format that we want it to be in later. Maps are what handle this work, making them extremely important for getting your work done!

There are two mapping functions which are frequently used:
    
    1. map():
    2. apply():

The function you pass to map() should expect a single value from the Series (a point value, in the above example), and return a transformed version of that value. map() returns a new Series where all the values have been transformed by your function.

mileage_mean = data.Mileage.mean()
mileage_mean

data.Mileage.map(lambda p:p - mileage_mean)
#It tansforms Mileage in data in form mileage - mean mileage

1. apply() transforms the whole DataFrame by calling a custom method on each row.
2. map() transforms the single series or column to say

def remean_Mileage(row):
    row.Mileage = row.Mileage - data.Mileage.mean()
    return row

data.apply(remean_Mileage, axis ='columns')

### If we had called data.apply() with axis='index', then instead of passing a function to transform each row, we would need to give a function to transform each column.

Note that map() and apply() return new, transformed Series and DataFrames, respectively. They don't modify the original data they're called on. 

More easy Way: of remean the Mileage:

data.Mileage - data.Mileage.mean()

data.Brand + '-' + data.City #Concat

data.Mileage.idxmax() #Will return the index of Max in the given column

data.Mileage.idxmin() #Will return the index of Min in the given column

# Question: Which is the best Mileage Car for Me?

data_id = data.Mileage.idxmax()
data_id

car = data.loc[data_id,'Brand']
car

data

# Question: Create a Series that counts how many times the word Maruti and Tata appears in our City?

Maruti_count= data.Brand.map(lambda x: 'Maruti' in x).sum()
Tata_count = data.Brand.map(lambda x: 'Tata' in x).sum()
total_count = pd.Series([Maruti_count,Tata_count],index = ['Maruti','Tata'])
total_count
