==> Pandas/Module 1/8. Data Types.ipynb <==
import pandas as pd
from pandas import DataFrame


data = pd.read_csv('new.csv')
data

#converter to make the data good:
def converter_na(cell):
    if cell == '':
        return 'Pune'
    return cell

def converter_date(cell):
    if cell =='':
        return 2011
    return cell

data = pd.read_csv('new.csv',converters= {
    'City' : converter_na
})

data

# Data Types:

The data type for a column in a DataFrame or a Series is known as the dtype.


data.Brand.dtype

data.Year.dtype

data.City.dtype

data.Mileage.dtype

data.dtypes #dtypes will return the data type of every column in our data set in general

One peculiarity to keep in mind (and on display very clearly here) is that columns consisting entirely of strings do not get their own type; they are instead given the object type

data.index.dtype #Datatype of the Index

It's possible to convert a column of one type into another wherever such a conversion makes sense by using the astype()

data.Mileage.astype('float64')

# Date Type:

data1 = pd.read_csv('weather.csv')
data1

type[data1.day]

df = pd.read_csv('weather.csv',parse_dates=['day'])
type(df.day[0])

df

df.set_index('day',inplace = True)
df

inplace = True will modify the original dataframe

Pandas also has some other data types called as Time series data and categorical data!!
