==> Pandas/Module 1/9. Handle the Missing Data.ipynb <==
import pandas as pd
import numpy as np
data = pd.read_csv('new.csv')
data

Entries missing values are given the value NaN, short for "Not a Number". For technical reasons these NaN values are always of the float64 dtype.

To select NaN entries you can use pd.isnull() (or its companion pd.notnull())

data1 = pd.isnull(data)
data1

data2 = pd.notnull(data)
data2

data[pd.isnull(data.Year)]

data[pd.notnull(data.City)]

Replacing missing values is a common operation. Pandas provides a really handy method for this problem: fillna(). fillna() provides a few different strategies for mitigating such data. For example, we can simply replace each NaN with an "Unknown"

data3 = data.fillna('Unknown')
data3

data4 = data.Year.fillna(0)
data4

or more specific:

new_df = data.fillna({'Year': 2000,
                     'City': 'Pune'})
new_df

# Carry forward and Carry Backward

newdf = data.fillna(method ='ffill')
newdf

newdf1 = data.fillna(method = 'bfill')
newdf1

newdf2 = data.fillna(method = 'bfill',axis ='columns')
newdf2
#here the cursor will go cell by cell row wise and copy the values likewise

newdf3 =data.fillna(method ='ffill', limit = 1)
newdf3
#limit will only carry forward one cell

newdf4 = data.interpolate()
newdf4

#interpolate took the middle value- Linear interpolation to come up with the values.
#see the interpolate methods in python documentation . by default it takes linear interpolation- middle value

#method = time will need date as index and then it will explore in time dimension. values close to dates will have values which are closer.

newdf5 = data.dropna() #it will drop the rows with na values
newdf5

#But, I dont want to drop all the rows. I want to drop rows only which have all the values as na then how?
newdf6 = data.dropna(how = 'all')
newdf6
#I will drop rows which has all na values. However our data set does not have such row

newdf7 = data.dropna(thresh = 1)
newdf7
#threshold = 1 means that I need atleas 1 valid value or else i will drop it


newdf8 = data.dropna(thresh = 4)
newdf8
#as row 2 had only three valid values so it is dropped

df = pd.read_csv('weather1.csv')
df

To handle the special values like 9999 use replace method:


df1 = df.replace([9999,99999],np.NaN)
df1

#also can replace value based on specific colum for that we need to give dictionary as input

df

new_df = df.replace({
    'temperature' : 99999,
    'windspeed': 9999,
    'city':999},
    np.NaN
)
new_df

new_df2 = df.replace('[A-Za-z]','',regex= True)
new_df2

dataframe = pd.DataFrame({
    'score': ['exceptional','good','average','poor','exceptional','average'],
    'student' : ['Amit','Ajit','Rony','Akansha','Mayuri','Poonam']
})
dataframe

dataframe.replace(['exceptional','good','average','poor'],['A','B','C','D'])

Using list can be pretty powerful in this case
