==> Pandas/Module 1/10. Rename and Combine.ipynb <==
import pandas as pd
from pandas import DataFrame
import numpy as np

data = pd.read_csv('new.csv')
data

# Renaming

 lets you change index names and/or column names

data.rename(columns = {'Kms Driven' : 'RideKm' })

rename() lets you rename index or column values by specifying a index or column keyword parameter, respectively. It supports a variety of input formats, but usually a Python dictionary is the most convenient.

data1= data.rename(index = { 0: 'first',
                            1 : 'Second'
})
data1

Both the row index and the column index can have their own name attribute. The complimentary rename_axis() method may be used to change these names. For example:

data1.rename_axis('ind',axis = 'rows').rename_axis('col',axis = 'columns')

# Combining

When performing operations on a dataset, we will sometimes need to combine different DataFrames and/or Series in non-trivial ways. Pandas has three core methods for doing this. In order of increasing complexity, these are concat(), join(), and merge()

# # Concat:

weather1 = DataFrame({
    'city' : ['Pune','Kolkata','Madras'],
    'Temperature' : [32,45,46],
    'humidity' : [80,60,78]
})
weather1

weather2 = DataFrame({
    'city' : ['Delhi','Bhopal','Imphal'],
    'Temperature' : [35,40,28],
    'humidity' : [50,65,90]
})
weather2

pd.concat([weather1,weather2])

pd.concat([weather1,weather2],ignore_index = True)

#pass Keys argument- helps associate keys with each dataframe

df = pd.concat([weather1,weather2],keys = ['one','two'])
df

df.loc['one'] #Helps in extracting more meaningfully:

tempdf =  DataFrame({
    'city' : ['Pune','Kolkata','Madras'],
    'Temperature' : [32,45,46]})
tempdf

winddf =  DataFrame({
    'city' : ['Pune','Kolkata','Madras'],
    'windspeed' : [100,200,300]})
winddf

pd.concat([tempdf,winddf])

#I want windspeed to appear as column and not to create new rows then what can I do?

pd.concat([tempdf,winddf],axis = 1)

winddf1 =  DataFrame({
    'city' : ['Madras','Kolkata','Pune'],
    'windspeed' : [100,200,300]})
'''Here the city are in different order then?''' 
winddf1


pd.concat([tempdf,winddf1],axis = 1)

'''to avoid the case like above we need to pass index when creating the df'''

winddf =  DataFrame({
    'city' : ['Pune','Kolkata','Madras'],
    'windspeed' : [100,200,300]},
index = [0,1,2])
winddf

tempdf =  DataFrame({
    'city' : ['Madras','Kolkata','Pune'],
    'Temperature' : [32,45,46]},
index = [2,1,0])
tempdf

pd.concat([tempdf,winddf],axis = 1)

'''when creating the df it is important that we use index and that helps us in aligning the data properly when we concatenate'''

s= pd.Series(['Humid','Dry','Rainy'],name = 'event')
s

pd.concat([tempdf,s], axis = 1)

 Now here one important problem that we faced is that everytime we had to concatenate two df we had to specify index, wouldnt it be best that the program does that on its own- Then we need to learn one important function called merge()

# MERGE

pd.merge(tempdf,winddf, on='city') 
#it is looking for city and then joining them

tempdf1 = DataFrame({
    'city' : ['Madras','Kolkata','Pune','Buldana'],
    'Temperature' : [32,45,46,50]})

pd.merge(tempdf1,winddf,on = 'city') 
#as we see the merge does the intersection and does not return the union as buldana is missing

pd.merge(tempdf1,winddf,on = 'city',how ='inner') #inner by default 


pd.merge(tempdf1,winddf,on = 'city', how ='left')  #left join

pd.merge(tempdf1,winddf,on = 'city', how ='right')  #right join

pd.merge(tempdf1,winddf,on = 'city', how ='left', indicator = True)  #inidicator set to true will indicate from where did the data come from

pd.merge(weather1,weather2,on = 'city',how = 'outer', suffixes = ('_left','_right'), indicator = True )

# Join()

left = tempdf.set_index('city')
right = winddf.set_index('city')
left.join(right)
