==> Pandas/Module 1/11. Pivot Table.ipynb <==
import pandas as pd
from pandas import DataFrame, Series
import numpy as np
data = pd.read_csv('weather.csv')
data

data.pivot(index ='day', columns = 'city')

data.pivot(index ='day', columns = 'city', values = 'temperature')

df = pd.read_csv('weather2.csv')
df

df.pivot(index='city', columns ='day')

df.pivot_table(index='city', columns ='day')  #gives the average if there are two values with same date

df.pivot_table(index='city', columns ='day', aggfunc = 'sum') #since column is day it will give sum of same day

df.pivot_table(index='city', columns ='day', aggfunc = 'count')

df.pivot_table(index='city', columns ='day', aggfunc = 'mean') #default

df.pivot_table(index='city', columns ='day', margins = 'True')

df['day']=pd.to_datetime(df['day'])
df.pivot_table(index = pd.Grouper(freq='M',key='day'), columns ='city')
