==> Pandas/Module 1/12. Stacking and Unstacking.ipynb <==
import pandas as pd

data = pd.read_excel('stocks.xlsx',header = [0,1],index_col = 0)
data

data.stack()

#level:

df_stacked = data.stack(level=0)
df_stacked

df_stacked.unstack()

df2 = pd.read_excel('stocks_3_levels.xlsx',header = [0,1,2], index_col = 0)
df2

df2.stack(level=0)

df2.stack(level=1)

df2.stack(level=2)
