==> Pandas/Module 1/13. CrossTab.ipynb <==
useful in case we need to form and display the contingency table

import pandas as pd

data = pd.read_csv("/home/meta-sapien/ineuron/Pandas/survey.csv")
data

pd.crosstab(data.Nationality,data.Handedness)

pd.crosstab(data.Sex,data.Handedness, margins = True)

pd.crosstab(data.Sex,[data.Nationality,data.Handedness], margins = True)

pd.crosstab([data.Sex],[data.Handedness],normalize = 'index')

# Q Find average age of female and Male who are R and L handed respectively::

pd.crosstab([data.Sex],[data.Handedness])

import numpy as np

pd.crosstab([data.Sex],[data.Handedness],values = data.Age,aggfunc = np.average)
